+++ src/components/header/BackShareHeader.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { View, SafeAreaView } from "react-native";
 import BackIcon from "../../assets/images/icon/Back.svg";
 import ShareIcon from "../../assets/images/icon/Share.svg";
-import { styles } from "./MainHeaderStyle";
+import { styles } from "../../styles/header/MainHeaderStyle";
 import { useNavigation } from "@react-navigation/native";
 
 export default function BackShareHeader() {
