+++ src/components/header/MainHeader.tsx
@@ -3,7 +3,7 @@ import { View, SafeAreaView } from "react-native";
 import LogoIcon from "../../assets/images/icon/Logo.svg";
 import AlertIcon from "../../assets/images/icon/Alert.svg";
 import AlertVariantIcon from "../../assets/images/icon/AlertVariant.svg";
-import { styles } from "./MainHeaderStyle";
+import { styles } from "../../styles/header/MainHeaderStyle";
 
 export default function MainHeader({ navigation }: { navigation: any }) {
   const [isAlert, setIsAlert] = useState(false);
