+++ src/components/libary/CarouselItem.tsx
@@ -6,7 +6,7 @@ import {
   Animated,
   Pressable,
 } from "react-native";
-import { styles } from "./CarouselItemStyle";
+import { styles } from "../../styles/library/CarouselItemStyle";
 import { TBookCategory } from "../../types/book";
 import BookItem from "./BookItem";
 import { dummyList } from "../../assets/data/dummyBookCarouseList";
