+++ src/styles/detail/DetailPageStyle.ts
@@ -1,12 +1,12 @@
 import { StyleSheet } from "react-native";
-import { Color } from "../../styles/Theme";
+import { Color } from "../Theme";
 
 export const styles = StyleSheet.create({
   container: {
     flex: 1,
     backgroundColor: Color.Secondary,
   },
-  wrap: {
+  scrollWrap: {
     flex: 1,
     alignItems: "center",
   },
