+++ src/styles/header/BackShareHeaderStyle.ts
@@ -1,12 +1,11 @@
 import { StyleSheet } from "react-native";
-import { Color } from "../../styles/Theme";
+import { Color } from "../Theme";
 
 export const styles = StyleSheet.create({
   container: {
     flexDirection: "row",
     justifyContent: "space-between",
-    paddingVertical: 14,
-    paddingHorizontal: 16,
+    padding: 16,
     alignItems: "center",
     backgroundColor: Color.Secondary,
   },
