+++ src/styles/header/MainHeaderStyle.ts
@@ -1,11 +1,12 @@
 import { StyleSheet } from "react-native";
-import { Color } from "../../styles/Theme";
+import { Color } from "../Theme";
 
 export const styles = StyleSheet.create({
   container: {
     flexDirection: "row",
     justifyContent: "space-between",
-    padding: 16,
+    paddingVertical: 14,
+    paddingHorizontal: 16,
     alignItems: "center",
     backgroundColor: Color.Secondary,
   },
