+++ src/styles/library/LibraryPageStyle.ts
@@ -1,5 +1,5 @@
 import { StyleSheet } from "react-native";
-import { Color } from "../../styles/Theme";
+import { Color } from "../Theme";
 
 export const styles = StyleSheet.create({
   container: {
