+++ src/styles/library/OnboardingOverlayStyle.ts
@@ -1,5 +1,5 @@
 import { StyleSheet } from "react-native";
-import { Color, Font } from "../../styles/Theme";
+import { Color, Font } from "../Theme";
 
 export const styles = StyleSheet.create({
   overlayTouch: {
