+++ src/styles/modal/ModalStyle.ts
